Remove AssetMapper, MinifyBundle, TailwindBundle, move to Reprise, install pnpm dependencies
