+++ .gitignore
@@ -1,7 +1,6 @@
 .cache
 /public/sitemap.*
 /Makefile.local
-/node_modules/
 
 ###> symfony/framework-bundle ###
 /.env.local
@@ -13,11 +12,6 @@
 /vendor/
 ###< symfony/framework-bundle ###
 
-###> symfony/asset-mapper ###
-/public/assets/
-/assets/vendor/
-###< symfony/asset-mapper ###
-
 ###> phpstan/phpstan ###
 phpstan.neon
 ###< phpstan/phpstan ###
@@ -27,10 +21,6 @@ phpstan.neon
 /.phpunit.cache/
 ###< phpunit/phpunit ###
 
-###> sensiolabs/minify-bundle ###
-/var/minify/
-###< sensiolabs/minify-bundle ###
-
 ###> kocal/oxc-bundle ###
 /bin/oxfmt
 /bin/oxlint
@@ -39,3 +29,8 @@ phpstan.neon
 ###> deptrac/deptrac ###
 /.deptrac.cache
 ###< deptrac/deptrac ###
+
+###> symfony/reprise ###
+/node_modules/
+/public/build/
+###< symfony/reprise ###
