+++ assets/admin.js
@@ -1,3 +1,3 @@
-import { startStimulusApp } from "@symfony/stimulus-bundle";
+import { startStimulusApp } from "@symfony/reprise/stimulus";
 
 const app = startStimulusApp();
