+++ assets/app.js
@@ -1,4 +1,4 @@
-import { startStimulusApp } from "@symfony/stimulus-bundle";
+import { startStimulusApp } from "@symfony/reprise/stimulus";
 import "./styles/app.css";
 
 const app = startStimulusApp();
