+++ assets/styles/app.css
@@ -27,7 +27,7 @@
         font-display: swap;
         font-style: normal;
         font-named-instance: "Regular";
-        src: url('../vendor/inter-ui/variable/InterVariable.woff2') format("woff2")
+        src: url('inter-ui/variable/InterVariable.woff2') format("woff2")
     }
     @font-face {
         font-family: Inter var;
@@ -35,7 +35,7 @@
         font-display: swap;
         font-style: italic;
         font-named-instance: "Regular";
-        src: url('../vendor/inter-ui/variable/InterVariable-Italic.woff2') format("woff2")
+        src: url('inter-ui/variable/InterVariable-Italic.woff2') format("woff2")
     }
 }
