+++ composer.json
@@ -24,9 +24,7 @@
         "phpstan/phpdoc-parser": "^2.3.2",
         "predis/predis": "^3.4",
         "presta/sitemap-bundle": "^4.3",
-        "sensiolabs/minify-bundle": "^1.2",
         "symfony/asset": "8.1.*",
-        "symfony/asset-mapper": "8.1.*",
         "symfony/clock": "8.1.*",
         "symfony/console": "8.1.*",
         "symfony/dotenv": "8.1.*",
@@ -40,6 +38,7 @@
         "symfony/property-access": "8.1.*",
         "symfony/property-info": "8.1.*",
         "symfony/rate-limiter": "8.1.*",
+        "symfony/reprise": "^0.1",
         "symfony/runtime": "8.1.*",
         "symfony/serializer": "8.1.*",
         "symfony/stimulus-bundle": "^3.2",
@@ -53,7 +52,6 @@
         "symfony/validator": "8.1.*",
         "symfony/web-link": "8.1.*",
         "symfony/yaml": "8.1.*",
-        "symfonycasts/tailwind-bundle": "^0.13.0",
         "twig/extra-bundle": "^3.24.0",
         "twig/html-extra": "^3.24.0",
         "twig/intl-extra": "^3.26.0",
@@ -118,9 +116,7 @@
     "scripts": {
         "auto-scripts": {
             "cache:clear": "symfony-cmd",
-            "assets:install %PUBLIC_DIR%": "symfony-cmd",
-            "importmap:install": "symfony-cmd",
-            "tailwind:build --minify": "symfony-cmd"
+            "assets:install %PUBLIC_DIR%": "symfony-cmd"
         },
         "post-install-cmd": ["@auto-scripts"],
         "post-update-cmd": ["@auto-scripts"]
