--- config/packages/asset_mapper.yaml
@@ -1,12 +1,0 @@
-framework:
-    asset_mapper:
-        importmap_polyfill: false
-        # The paths to make available to the asset mapper.
-        paths:
-            - assets
-        missing_import_mode: strict
-
-when@prod:
-    framework:
-        asset_mapper:
-            missing_import_mode: warn
