--- config/packages/sensiolabs_minify.yaml
@@ -1,19 +1,0 @@
-sensiolabs_minify:
-    asset_mapper:
-        # Exclude already minified assets
-        ignore_paths:
-            - '*.min.js'
-            - '*.min.css'
-    minify:
-        # Use local binary if available
-        # binary_path: 'auto'
-
-        # Specify local binary path
-        # binary_path: '/usr/bin/minify'
-
-        # Download binary from GitHub
-        download_binary: true
-
-when@test:
-    sensiolabs_minify:
-        asset_mapper: false
