--- config/packages/symfonycasts_tailwind.yaml
@@ -1,2 +1,0 @@
-symfonycasts_tailwind:
-    binary_version: v4.3.0
