+++ package.json
@@ -1,5 +1,44 @@
 {
-  "name": "hugo-alliau-me",
   "private": true,
-  "packageManager": "pnpm@11.12.0"
+  "packageManager": "pnpm@11.12.0",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build"
+  },
+  "dependencies": {
+    "@codemirror/commands": "^6.10.4",
+    "@codemirror/lang-css": "^6.3.1",
+    "@codemirror/lang-html": "^6.4.11",
+    "@codemirror/lang-javascript": "^6.2.5",
+    "@codemirror/lang-json": "^6.0.2",
+    "@codemirror/lang-markdown": "^6.5.0",
+    "@codemirror/lang-php": "^6.0.2",
+    "@codemirror/lang-sass": "^6.0.2",
+    "@codemirror/lang-vue": "^0.1.3",
+    "@codemirror/lang-yaml": "^6.1.3",
+    "@codemirror/language": "^6.12.4",
+    "@codemirror/legacy-modes": "^6.5.3",
+    "@codemirror/state": "^6.7.1",
+    "@codemirror/view": "^6.43.6",
+    "@fsegurai/codemirror-theme-github-dark": "^6.2.5",
+    "@fsegurai/codemirror-theme-github-light": "^6.2.5",
+    "@hotwired/stimulus": "^3.2.2",
+    "@hotwired/turbo": "^7.1.0 || ^8.0",
+    "@lezer/common": "^1.5.2",
+    "@lezer/highlight": "^1.2.3",
+    "@lezer/lr": "^1.4.10",
+    "@symfony/ux-leaflet-map": "file:vendor/symfony/ux-leaflet-map/assets",
+    "@symfony/ux-turbo": "file:vendor/symfony/ux-turbo/assets",
+    "codemirror": "^6.0.2",
+    "inter-ui": "^4.1.1",
+    "leaflet": "^1.9.4",
+    "stimulus-use": "^0.53.0"
+  },
+  "devDependencies": {
+    "@symfony/reprise": "^0.1.0",
+    "@symfony/stimulus-bridge": "^3.2.0 || ^4.0.0",
+    "@tailwindcss/vite": "^4.3.2",
+    "tailwindcss": "^4.3.2",
+    "vite": "^8.1.4"
+  }
 }
