+++ pnpm-workspace.yaml
@@ -1,0 +1,2 @@
+minimumReleaseAgeExclude:
+  - '@symfony/reprise@0.1.0'
