+++ symfony.lock
@@ -108,33 +108,6 @@
     "presta/sitemap-bundle": {
         "version": "v4.1.1"
     },
-    "sensiolabs/minify-bundle": {
-        "version": "0.9",
-        "recipe": {
-            "repo": "github.com/symfony/recipes-contrib",
-            "branch": "main",
-            "version": "0.9",
-            "ref": "e2583c294ba6615da8bedff18a8dbf55e4562a6d"
-        },
-        "files": [
-            "config/packages/sensiolabs_minify.yaml"
-        ]
-    },
-    "symfony/asset-mapper": {
-        "version": "7.3",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "main",
-            "version": "6.4",
-            "ref": "5ad1308aa756d58f999ffbe1540d1189f5d7d14a"
-        },
-        "files": [
-            "assets/app.js",
-            "assets/styles/app.css",
-            "config/packages/asset_mapper.yaml",
-            "importmap.php"
-        ]
-    },
     "symfony/console": {
         "version": "7.1",
         "recipe": {
@@ -247,6 +220,20 @@
         },
         "files": [
             "config/packages/property_info.yaml"
+        ]
+    },
+    "symfony/reprise": {
+        "version": "0.1",
+        "recipe": {
+            "repo": "github.com/symfony/recipes",
+            "branch": "main",
+            "version": "0.1",
+            "ref": "912ce6c6bd6b81afac04525159bfcceb08fdd9fe"
+        },
+        "files": [
+            "assets/app.js",
+            "assets/styles/app.css",
+            "config/packages/reprise.yaml"
         ]
     },
     "symfony/routing": {
@@ -407,9 +394,6 @@
             "config/packages/web_profiler.yaml",
             "config/routes/web_profiler.yaml"
         ]
-    },
-    "symfonycasts/tailwind-bundle": {
-        "version": "v0.5.1"
     },
     "twig/extra-bundle": {
         "version": "v3.8.0"
