+++ templates/components/AppFooter.html.twig
@@ -6,7 +6,7 @@
         <div>
             <div class="flex justify-center gap-x-4 px-2">
                 <img
-                    src="{{ asset('images/hugo-alliaume.jpeg') }}"
+                    src="{{ asset('build/images/hugo-alliaume.jpeg') }}"
                     width="92"
                     height="92"
                     decoding="async"
