+++ templates/cv/home.html.twig
@@ -12,7 +12,7 @@
             <twig:Card class="p-2 sm:p-3 flex print:shadow-none print:border">
                 <div class="flex-1 flex flex-col justify-between gap-2">
                     <img
-                        src="{{ asset('images/hugo-alliaume.jpeg') }}"
+                        src="{{ asset('build/images/hugo-alliaume.jpeg') }}"
                         width="169"
                         height="169"
                         decoding="async"
